extract jarSearchableOptions task to separate class
